fix(opencode): stop leaking helper exports that crash OpenCode v1.17.4
The plugin entry module exported two test-only helpers
(`applyLoreProviderConfig`, `probeGateway`) alongside `LorePlugin`.
OpenCode's legacy plugin loader treats every function exported from a
plugin module as a plugin instance and invokes it (`getServerPlugin` /
`getLegacyPlugins`). Those helpers don't return a `Hooks` object, so
`undefined` was pushed into the host's hooks array. The first hook
dispatch then crashed the host:

    undefined is not an object (evaluating 'A.event')

(`A` is the minified `hook`; the `?.` guards `.event` but not an
undefined `hook` element). The same defect surfaced on the `config`
and `dispose` dispatch loops (`j.config` / `j.dispose`), and the
`event` variant became a fatal "Unexpected server error".

Fix: move the helpers into `src/internal.ts` so the plugin entry
exposes only `LorePlugin` (+ same-ref default). Tests import the
helpers from the internal module. Adds a regression test that
simulates `getLegacyPlugins` dedup and asserts the entry exposes
exactly one plugin function.

Verified end-to-end against opencode-ai@1.17.4: plugin loads and the
event/config/dispose dispatch errors are gone.
