+++ packages/opencode/src/index.ts
@@ -5,6 +5,12 @@ import {
   discoverWorkspaceRoot,
   installFetchInterceptor,
 } from "@loreai/core";
+// Helpers live in a separate module so they are NOT re-exported from this
+// plugin entry. OpenCode's legacy plugin loader invokes every function
+// exported from the entry module as a plugin; leaking these helpers pushed
+// `undefined` into the host hooks array and crashed it on event dispatch
+// (`undefined is not an object (evaluating 'A.event')`). See ./internal.ts.
+import { applyLoreProviderConfig, probeGateway } from "./internal";
 
 /**
  * Lore plugin for OpenCode — transparent LLM proxy routing.
@@ -22,70 +28,6 @@ import {
 
 /** Default ports to probe when looking for a running gateway (must match gateway defaults). */
 const KNOWN_GATEWAY_PORTS = [3207, 5673];
-
-/**
- * Pin every opencode provider's `options.baseURL` to the Lore gateway.
- * Without this, opencode can derive the Anthropic baseURL from
- * `OPENAI_BASE_URL` (stripping `/v1`), sending the SDK to
- * `http://host/messages` (no /v1) — the gateway only routes `/v1/messages`,
- * and the fetch interceptor skips 127.0.0.1 to avoid loops, so the call
- * lands as a bare `/messages` 404. Worse, the `OPENAI_BASE_URL` /
- * `ANTHROPIC_BASE_URL` env vars are bypassed by opencode's `resolveSDK()`
- * (it always passes `options.baseURL` to the @ai-sdk factory, and the
- * @ai-sdk `loadOptionalSetting()` only consults the env var when the
- * factory receives an undefined `baseURL`). Every other @ai-sdk provider
- * (google, mistral, groq, cohere, xai, perplexity, togetherai, vercel,
- * alibaba, deepinfra, gateway, openrouter, cerebras, etc.) has NO
- * baseURL env var at all. Iterating over `cfg.provider` is the only
- * universal lever.
- *
- * Deep-merges per-provider so user-set keys under `provider.<id>` (custom
- * headers, model overrides, etc.) are preserved.
- *
- * Exported for direct testing — the config hook delegates here so unit tests
- * can verify the merge logic without spinning up a real gateway (the
- * surrounding `LorePlugin` skips gateway start in `NODE_ENV=test`).
- */
-export function applyLoreProviderConfig(
-  cfg: Record<string, unknown>,
-  gatewayBase: string,
-): void {
-  if (!gatewayBase) return;
-  const baseUrl = `${gatewayBase}/v1`;
-  const existingProvider = (cfg.provider ?? {}) as Record<string, unknown>;
-  const pinned: Record<string, unknown> = {};
-  for (const [id, provider] of Object.entries(existingProvider)) {
-    if (!provider || typeof provider !== "object") continue;
-    const p = provider as Record<string, unknown>;
-    const existingOptions = (p.options ?? {}) as Record<string, unknown>;
-    pinned[id] = {
-      ...p,
-      options: { ...existingOptions, baseURL: baseUrl },
-    };
-  }
-  if (Object.keys(pinned).length > 0) {
-    cfg.provider = { ...existingProvider, ...pinned };
-  }
-}
-
-/**
- * Check if the Lore gateway is reachable at the given base URL.
- * Short timeout so this doesn't delay OpenCode startup noticeably.
- */
-export async function probeGateway(
-  baseURL: string,
-  timeoutMs = 1500,
-): Promise<boolean> {
-  try {
-    const controller = new AbortController();
-    const timer = setTimeout(() => controller.abort(), timeoutMs);
-    const res = await fetch(`${baseURL}/health`, { signal: controller.signal });
-    clearTimeout(timer);
-    return res.ok;
-  } catch {
-    return false;
-  }
-}
 
 /**
  * Resolve the gateway URL by probing known ports and reading the port file.
